third_party/WebKit/LayoutTests/fast/events/touch/touch-user-gesture-cross-origin-expected.txt



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
Target
Test user gesture behavior during touch events.
touchstart should not be a user gesture
PASS openedPopup is false
touchmove should not be a user gesture
PASS openedPopup is false
touchend should not be a user gesture if it's moved beyond the slop region
PASS openedPopup is false
touchend should be a user gesture when it occurs as part of a tap
PASS openedPopup is true
touchmove and touchend should not be a user gesture when it occurs as part of a drag without scrolling
PASS openedPopup is false
PASS openedPopup is false
PASS successfullyParsed is true

TEST COMPLETE
